fix(ci): split path-extraction tests out of tests.rs; unlink private match_glob
Two gates the local sweep did not cover went red on d73ac71:

- The 1500-line ratchet: the new turn_path_tokens cases grew
  stella-cli/src/memory/tests.rs past its grandfathered ceiling
  (1570 > 1526). The path-anchor/path-token cluster moves to its own
  path_token_tests.rs module, putting tests.rs at 1498 — under the
  plain limit, so its baseline exemption is retired too.
- cargo doc -D warnings: the applies_to module doc intra-doc-linked
  match_glob, which is crate-private; public docs reject that. Plain
  code formatting keeps the reference without the link.
